feat: change length type to size_t

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: loda-sil <loda-sil@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/23 19:28:47 by loda-sil          #+#    #+#             */
-/*   Updated: 2025/07/24 14:45:29 by loda-sil         ###   ########.fr       */
+/*   Updated: 2025/07/29 12:39:29 by loda-sil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	length;
+	size_t	length;
 
 	length = 0;
 	while (s[length])
